+++ app/src/main/AndroidManifest.xml
@@ -18,19 +18,19 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".SignUpActivity"
+            android:name=".feature.signup.SignUpActivity"
             android:windowSoftInputMode="stateAlwaysHidden|adjustResize"
             android:exported="false" />
         <activity
-            android:name=".ProfileActivity"
+            android:name=".feature.profile.ProfileActivity"
             android:windowSoftInputMode="stateAlwaysHidden|adjustResize"
             android:exported="false" />
         <activity
-            android:name=".HomeActivity"
+            android:name=".feature.home.HomeActivity"
             android:windowSoftInputMode="stateAlwaysHidden|adjustResize"
             android:exported="false"/>
         <activity
-            android:name=".MainActivity"
+            android:name=".feature.main.MainActivity"
             android:windowSoftInputMode="stateAlwaysHidden|adjustResize"
             android:exported="true">
             <intent-filter>
