--- app/src/main/java/org/android/go/sopt/GalleryViewModel.kt
@@ -1,42 +1,0 @@
-package org.android.go.sopt
-
-import android.util.Log
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import org.android.go.sopt.data.remote.ServicePool.imageService
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
-
-class GalleryViewModel : ViewModel() {
-    private val _image = MutableLiveData<ContentUriRequestBody>()
-    val image: LiveData<ContentUriRequestBody>
-    get() = _image
-
-
-
-
-    fun setRequestBody(requestBody: ContentUriRequestBody) {
-        _image.value = requestBody
-    }
-
-    fun uploadProfileImage() {
-        if (image.value == null) { /* 아직 사진이 등록되지 않았다는 로직 처리 */
-        } else {
-            imageService.uploadImage(image.value!!.toFormData()).enqueue(
-                object : Callback<Unit> {
-                    override fun onResponse(call: Call<Unit>, response: Response<Unit>) {
-                        if (response.isSuccessful)
-                            Log.e("emjay", "success !!!")
-                    }
-
-                    override fun onFailure(call: Call<Unit>, t: Throwable) {
-                        TODO("Not yet implemented")
-                    }
-
-                }
-            )
-        }
-    }
-}
