+++ app/src/main/java/org/android/go/sopt/data/datasource/model/SignUPDTO.kt
@@ -1,0 +1,34 @@
+package org.android.go.sopt.data.datasource.model
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RequestSignUpDto(
+    @SerialName("id")
+    val id: String,
+    @SerialName("password")
+    val password: String,
+    @SerialName("name")
+    val name: String,
+    @SerialName("skill")
+    val skill: String,
+)
+
+@Serializable
+data class ResponseSignUpDto(
+    @SerialName("status")
+    val status: Int,
+    @SerialName("message")
+    val message: String,
+    @SerialName("data")
+    val data: SignUpData,
+) {
+    @Serializable
+    data class SignUpData(
+        @SerialName("name")
+        val name: String,
+        @SerialName("skill")
+        val skill: String,
+    )
+}
