+++ app/src/main/java/org/android/go/sopt/data/datasource/remote/AuthInterceptor.kt
@@ -1,0 +1,16 @@
+package org.android.go.sopt.data.datasource.remote
+
+
+import okhttp3.Interceptor
+import okhttp3.Response
+
+class AuthInterceptor : Interceptor {
+    override fun intercept(chain: Interceptor.Chain) : Response {
+        val originalRequest = chain.request()
+        val headerRequest = originalRequest.newBuilder()
+            .header("Authorization", "")
+            .build()
+
+        return chain.proceed(headerRequest)
+    }
+}
