--- app/src/main/java/org/android/go/sopt/data/remote/AuthInterceptor.kt
@@ -1,16 +1,0 @@
-package org.android.go.sopt.data.remote
-
-
-import okhttp3.Interceptor
-import okhttp3.Response
-
-class AuthInterceptor : Interceptor {
-    override fun intercept(chain: Interceptor.Chain) : Response {
-        val originalRequest = chain.request()
-        val headerRequest = originalRequest.newBuilder()
-            .header("Authorization", "")
-            .build()
-
-        return chain.proceed(headerRequest)
-    }
-}
