--- app/src/main/java/org/android/go/sopt/data/remote/model/LogInDTO.kt
@@ -1,32 +1,0 @@
-package org.android.go.sopt.data.remote.model
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RequestLogInDto(
-    @SerialName("id")
-    val id: String,
-    @SerialName("password")
-    val password: String
-)
-
-@Serializable
-data class ResponseLogInDto(
-    @SerialName("status")
-    val status: Int,
-    @SerialName("message")
-    val message: String,
-    @SerialName("data")
-    val data: SignUpData,
-) {
-    @Serializable
-    data class SignUpData(
-        @SerialName("id")
-        val id: String,
-        @SerialName("name")
-        val name: String,
-        @SerialName("skill")
-        val skill: String,
-    )
-}
