+++ app/src/main/java/org/android/go/sopt/feature/main/MainActivity.kt
@@ -1,0 +1,48 @@
+package org.android.go.sopt.feature.main
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.view.MotionEvent
+import android.view.inputmethod.InputMethodManager
+import androidx.activity.viewModels
+import org.android.go.sopt.feature.home.HomeActivity
+import org.android.go.sopt.feature.signup.SignUpActivity
+import org.android.go.sopt.databinding.ActivityMainBinding
+
+class MainActivity : AppCompatActivity() {
+    private val binding by lazy {ActivityMainBinding.inflate(layoutInflater)}
+    private val viewModel by viewModels<LogInViewModel>()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(binding.root)
+
+
+        binding.btSi.setOnClickListener {
+            val intent = Intent(this, SignUpActivity::class.java)
+            startActivity(intent)
+        }
+
+
+        binding.btLog.setOnClickListener {
+            viewModel.logIn(
+                binding.etId.text.toString(),
+                binding.etPw.text.toString()
+            )
+        }
+
+        viewModel.logInResult.observe(this) { logInResult ->
+            startActivity(Intent(this, HomeActivity::class.java))
+        }
+    }
+
+
+    override fun dispatchTouchEvent(ev: MotionEvent): Boolean {
+        val imm: InputMethodManager =
+            getSystemService(INPUT_METHOD_SERVICE) as InputMethodManager
+        imm.hideSoftInputFromWindow(currentFocus?.windowToken, 0)
+        return super.dispatchTouchEvent(ev)
+    }
+
+}
