+++ app/src/main/java/org/android/go/sopt/util/ContentUriRequestBody.kt
@@ -1,0 +1,53 @@
+package org.android.go.sopt.util
+
+import android.content.Context
+import android.net.Uri
+import android.provider.MediaStore
+import okhttp3.MediaType
+import okhttp3.MediaType.Companion.toMediaTypeOrNull
+import okhttp3.MultipartBody
+import okhttp3.RequestBody
+import okio.BufferedSink
+import okio.source
+
+class ContentUriRequestBody(
+    context: Context,
+    private val uri: Uri,
+) : RequestBody() {
+    private val contentResolver = context.contentResolver
+
+    private var fileName = ""
+    private var size = -1L
+
+    init {
+        contentResolver.query(
+            uri,
+            arrayOf(MediaStore.Images.Media.SIZE, MediaStore.Images.Media.DISPLAY_NAME),
+            null,
+            null,
+            null,
+        )?.use { cursor ->
+            if (cursor.moveToFirst()) {
+                size = cursor.getLong(cursor.getColumnIndexOrThrow(MediaStore.Images.Media.SIZE))
+                fileName =
+                    cursor.getString(cursor.getColumnIndexOrThrow(MediaStore.Images.Media.DISPLAY_NAME))
+            }
+        }
+    }
+
+    fun getFileName() = fileName
+
+    override fun contentLength(): Long = size
+
+    override fun contentType(): MediaType? =
+        contentResolver.getType(uri)?.toMediaTypeOrNull()
+
+    override fun writeTo(sink: BufferedSink) {
+        contentResolver.openInputStream(uri)?.source()?.use { source ->
+            sink.writeAll(source)
+        }
+    }
+
+    // 얘는 SerialName을 붙여서 이름을 명시해주지 않았죠? 여기 밑에서 이름을 정해줍니다 !
+    fun toFormData() = MultipartBody.Part.createFormData("file", getFileName(), this)
+}
